fix(telemetry): apply review cleanup
